Refactored server to DDD structure
