+++ server/src/businesses/mutations/updateBusiness.ts
@@ -1,14 +1,15 @@
-import { businessSchema } from '../../utils/validateBusiness';
-import { BusinessModel } from '../../models/businessModel';
+import { businessSchema } from '../validate';
+import { BusinessModel } from '../model';
 import { NextFunction, Request, Response } from 'express';
 
-export async function updateBusiness(req: Request, res: Response, next: NextFunction): Promise<any> {
+export async function updateBusiness(req: Request, res: Response, next: NextFunction): Promise<void> {
   const { error } = businessSchema.validate(req.body);
   if (error) {
-    return res.status(400).json({
+    res.status(400).json({
       error: 'Validation failed',
       message: error.details.map((e) => e.message),
     });
+    return;
   }
 
   try {
