+++ server/src/businesses/types.ts
