+++ server/src/categories/types.ts
