+++ server/src/categories/validate.ts
