+++ server/src/users/generateToken.ts
